feat: add pixi_compute_network crate with HasDownloadClient
