+++ crates/pixi_command_dispatcher/src/install_pixi/ext.rs
@@ -14,8 +14,7 @@ use crate::BuildProfile;
 use crate::CommandDispatcherError;
 use crate::cache::markers::{SourceBuildArtifactsDir, SourceBuildWorkspacesDir};
 use crate::compute_data::{
-    HasAllowExecuteLinkScripts, HasAllowLinkOptions, HasDownloadClient, HasPackageCache,
-    HasPixiInstallReporter,
+    HasAllowExecuteLinkScripts, HasAllowLinkOptions, HasPackageCache, HasPixiInstallReporter,
 };
 use crate::install_pixi::{
     InstallPixiEnvironmentError, InstallPixiEnvironmentResult, InstallPixiEnvironmentSpec,
@@ -24,6 +23,7 @@ use crate::install_pixi::{
 use crate::keys::{ArtifactCache, SourceBuildKey, SourceBuildSpec, WorkspaceCache};
 use crate::reporter::PixiInstallReporter;
 use pixi_compute_cache_dirs::CacheDirsExt;
+use pixi_compute_network::HasDownloadClient;
 
 /// Extension trait on [`ComputeCtx`] that installs a pixi environment
 /// with source-build recursion routed through [`SourceBuildKey`].
